add 'dconf-update' utility
This utility scans /etc/dconf/db/ looking for directories named *.d.

For each directory name.d that it finds, it reads each file contained
within as a keyfile containing a list of keys to populate a dconf
database with.

That database is then written to 'name' (ie: the directory name with the
'.d' removed), the old file is invalidated, and a signal is sent over
the system DBus to indicate the change.

This tool facilitates updating of dconf databases by sysadmins who
prefer to follow the 'drop a text file here' approach to system
management.