Deal with incompatible changes in Vala
Vala 0.13 (leading toward 0.14) corrects an error in the binding for
GtkTreeModel.  Unfortunately, the correction is incompatible and there
is no way to have code that will work with both old and new Vala
compilers.

Rather than hard-depend on the unstable Vala release, just use two
separate declarations decided on #if VALA_0_14.