Support maybe types in dconf
Work around the fact that we can't send maybe types over D-Bus by
sending a blob of serialised GVariant data instead.

Abuse the fact that we're using an array to denote 'maybe' at the level
of the dconf protocol and fill that array with 2 items (which would not
otherwise be possible) to indicate this special case.