Remove 'set lock' support from dconf
The dconf service can not presently be run at the system level and it
doesn't make sense to support locks on user-level databases.  It also
became clear that most distributors and sysadmins would rather work
directly with text files anyway, so we supported that directly instead.

For this reason, 'set lock' support has never been properly implemented.

All the plumbing was added for it though, which means we have it
appearing on the API of the client library and documented in the help of
the commandline tool.  This is misleading, since these functions do
nothing at all (and actually contain bugs anyway since their
do-nothingness was never actually tested).

For now, we rip out these functions.  We can add them back later if we
decide to support this properly.