engine: fix checking for locks
The algorithm that checked for locks essentially answers the question of
"what is the highest level database that we should consider results
from?".  In the case of only one database, its answer to this question
could be out of range of the set of databases that we asked it to search
in.

Correct this problem by bounding the starting point of the search
accordingly.

Reported by Didier Roche.