dconf update: Read keyfiles in a defined order
This patch is a substantial cleanup of the 'dconf update' command.  The
biggest 'clean up' aspect is less dependence on catch-all 'throws'
clauses and better error messages for specific failures.

The cleanup also brings a new feature: keyfiles are now read in a
pre-defined order so that keys defined in higher-numbered files have
precedence over low-numbered files (ie: if the same key is in a file
'10_a' and '20_b' then '20_b' will "win").  This portion of the patch is
based on a patch from Josselin Mouette.

https://bugzilla.gnome.org/show_bug.cgi?id=670494