dconf update: don't warn on a key set twice
It is expected that users may have the same key set multiple times if
they use separate files in an 'override' type of situation.  In this
case the higher-numbered file wins and we should silently ignore the key
set in the lower-numbered file.

This silences the warning message that we previously issued in that
case.

https://bugzilla.gnome.org/show_bug.cgi?id=664288