Implement change signals
Support receiving and properly exposing change notifications.

This required some changes to improve the thread-safety of destroying a
DConfEngine.

It is possible that a signal would be arriving (in the worker thread) at
the exact instant that a DConfEngine was being destroyed (from the
finalize of the DConfClient or DConfSettingsBackend).  This could lead
to the object being accessed after it was finalized.

We can avoid this by using weak references and by being more careful
about when the DConfEngine is freed (by taking a ref to it in the signal
handler and releasing it when done).