engine: correct logic error in queue checking
The logic used for checking for changes in the pending and in-flight
queues was backwards meaning that the in-flight queue was only checked
if a match was found in the pending queue.  We rather want the in-flight
queue to be checked only if we have not found a match already.