clean up and factor out the 'shm' code
Remove the shm code from the engine and the service and put it in a
separate convenience library in shm/.

Remove the vestigial shmdir weirdness from the service (since shmdir is
now always relative to XDG_RUNTIME_DIR and has been for some time).

The purpose of this is so that dconf-engine can be properly unit-tested.
dconf-engine now has five points of contact with the world (excluding
the users of the engine themselves):

  - the DCONF_PROFILE environment variable

  - fopen() of profile files

  - shm

  - gvdb

  - dbus

The environment variable is quite easily controlled.  fopen() is
intercepted in the engine testcase with a interpose of the libc symbol.

With this commit now each of dbus, gvdb and shm are implemented in
separate utility modules that can be mocked from the testcases.