shm/: make some adjustments after research
After doing some research it has been discovered that ftruncate() isn't
sufficient for allocating space for a file.  It can create a sparse file
and you don't find out that the space doesn't exist until you get a
SIGBUS later upon trying to write to the mmap region.

posix_fallocate() isn't doing quite what we want either (due to a flaky
glibc emulation of it when the native syscall is unavailable -- which is
the case for tmpfs).

Use a workaround based on pwrite() instead.

Also: mmap() really can't fail here except in the cases where we would
also abort due to g_malloc() failing, so we should just abort too.