user source: change approach to shm
Use dconf_shm_is_flagged() instead of direct access.

Change the semantics of what happens when we fail to open the shm file.
We used to refuse to initialise at all in that case -- now we initialise
but will try again every time to reopen ourselves.

The error case will now be slower but closer to the correct behaviour.