massive Makefile reorganisation
Clean up the Makefiles and make them as similar as possible.

Move CFLAGS to a common point of definition and stop using -I so much.

Replace the 'dbus stub' with libdconf-mock.a in tests/.  Fill in some
stubs for future mock code for shm and gvdb (just to get things
compiling for now).