engine: fix dconf_engine_unref() thread safety
There was a very slim race condition in the implementation of
dconf_engine_unref().

In order to check if the engine should be freed (to avoid double-freeing
in two separate threads) the code was checking if the engine was still
in the global list of engines.

It is possible, however, that this thread could have unrefed the engine
at exactly the same time as another thread, that thread won the race and
freed the engine (removing it from the list) and then a third thread
created another engine at the same time and it happened to have the same
pointer address as the engine that was just freed.  In this case, the
first engine would still see "itself" in the global list and free again.

It's unlikely that this would ever happen in the real world but the
regression tests picked it up.

The new implementation should avoid that issue by not depending on the
engine finding itself in the list as part of the decision about if it
should be freed or not.