Add script to trim bogus lcov data
There is no (apparent) way to trim branch or line coverage data from
lcov output based on expression matching against the source file.  This
means that we get coverage results that tell us (for example) that we
failed to test lines like 'g_assert_not_reached ();' or didn't test all
branches of a g_assert().

Add a script to read the lcov output and remove line or branch coverage
data according to a list of symbols that are known to have incomplete
coverage when operating as expected.