dbus-1/: rewrite against the new engine
Delete the copy of the old engine code that was kept in dbus-1/.

Split the parts that interact with libdbus-1 into a separate file called
dconf-libdbus-1.c and create a static library for it (to facilitate
testing).

Adjust the remaining code (which is now the user-facing DConfDBusClient
API) to the new engine.