tests/: test the libdbus-1 DBus backend
Add a testcase for the libdbus-1 DBus backend.  It uses the existing
testcase code and undergoes the same tests as the two other backends.

There is one exception: we do not test for failure to connect to D-Bus
because the D-Bus connections are passed up-front for the libdbus-1
case, as arguments to dconf_dbus_client_new().