engine/: fix refcounting error
The new supposedly-threadsafe implementation of refcounting in the
engine still had a bug in it: in the event that we were going to remove
the last reference and we discovered that the refcount was not 1 after
taking the lock, we would return (under the assumption that someone else
had increased the count).

This is the wrong thing to do.  No matter what, we need to drop our
reference -- instead of return, we should goto again.