DConfEngineSource: return FALSE from NULL refresh
In the case that we call dconf_engine_source_refresh() on a source that
had a NULL database and the result is that we still have a NULL
database, return FALSE.

This will prevent the unnecessary bumping of the state counter when
there was really no change.  This happens in the case of a missing
system database file.