tests/: test reading from various profile setups
Add a testcase that tests the engine by reading from an exhaustive
combination of different profile types and states (missing databases,
empty databases, databases with values, databases with locks, etc).

Among other things, this makes sure the lockdown logic is sane.

This is the testcase that caught the bug fixed in the last commit
(listing with a missing database file).