engine: improve robustness of profile parsing
Checking for 'u' or 's' is really insanely silly from a robustness
standpoint.  Ensure that we properly have "user-db:" or "system-db:" and
that a non-empty database name is given, warning if not.

This was specifically causing annoying problems with the profile file
that gdm was installing, so add a copy of that file to our testcases.