Add man pages
It is a nice thing if one can learn something about the running
processes by looking up their man pages. This commit adds man pages
for dconf-service, dconf-editor and the dconf utility in section 1.
It also adds an overview man page in section 7. The contents of
the overview are mostly taken from the current README. All these
man pages are also included in the API docs, in a 'Programs'
section.

https://bugzilla.gnome.org/show_bug.cgi?id=680602