service: rewrite
Rewrite the dconf-service using gdbus-codegen and generally cleaning
things up a lot.  The DConfWriter class can now be reasonably subclassed
to create more complex types of dconf databases (such as ones that
are stored in the local runtime dir and synced up with an NFS home
directory).

Keep a cache of the keys in the database (instead of re-reading it every
time we try to make a change).

Drop support for the old D-Bus interface (now that we are two stable
releases since it was used).

Modify the commandline tool for 'dconf blame' to call the new interface.