DConfChangeset: add "database mode" and change()
Add a new mode for DConfChangeset to represent the entire contents of a
dconf database.

Also add a dconf_changeset_change() call to either merge changesets or
apply a changeset to a database-mode changeset.

Doing this allows us to have only one instance of the algorithm to deal
with resets (ie: when resetting '/a/' we must reset all keys starting
with '/a/').