service: use database-mode DConfChangeset
Port the DConfWriter to use database-mode changesets to represent the
contents of the database.

This allows us to drop our own copy of the reset algorithm.

Also, fix a bug whereby we tried to acquire a list of items from a
NULL gvdb table (in the case where the database does not already exist).