engine: add a new database type: "service-db"
service-db databases are based on a gvdb in the user runtime dir
maintained by the dconf-service.

On startup, the client will send an Init message to the service to
ensure that the database is properly initialised.  After that, things
are pretty much the same except that the values are read from the file
in the runtime dir instead of from the home directory.

We also drop the "shm" signalling mechanism for this case and instead
use the same invalidation trick that system databases use (ie:
overwriting the old gvdb header after writing the new file).