keyfile: many improvements
Store the keyfile in ~/.config/dconf/$(name).txt instead of
~/.config/dconf-keyfile/name.  Adjust the base writer not to pick up
filenames with dots in them (so we will skip the .txt files the keyfile
writer puts into the same directory).

Add file change monitoring and reload the file if anyone changes it.

Fix a crasher on the first write of a file if it does not yet exist.

Add proper finalize handling.