build: "&>" is a advanced shell extension. Use > and 2>&1 instead.
dconf autogen.sh uses /bin/sh as shebang which on debian bind to dash.
Dash has no support for "&>" redirection (both stdout and stderr). Thus
it interpret the command as detach and redirect stdout.
Switch to explicitely redirect stdout and stderr :
"> /dev/null 2>&1" for which "&>" is a shorthand to fix the build
failure.

https://bugzilla.gnome.org/show_bug.cgi?id=692186