writer: ignore empty changesets
When emitting change signals, we call dconf_changeset_describe() to
enumerate the keys to send the signal for.

When the changeset is empty, this function returns NULL for the path
vector.  We pass that NULL into the signal emitter and the service
crashes.

We can avoid this situation by refusing to handle empty changesets in
the first place.