DConfChangeset: expose concept of "sealing"
DConfChangeset is a partially threadsafe type.

when first created, it is mutable and can only be used from one thread.
After it is filled in, the intention is that it can be shared between
threads as long as it isn't changed.

Previously, this transition was made when dconf_changeset_describe() was
called.  After that, it was not possible to make any more changes.

Formalise and document this concept and add an explicit call for it:
dconf_changeset_seal().

https://bugzilla.gnome.org/show_bug.cgi?id=703073