engine: issue warnings once per source
On failure to open a gvdb file in /etc/dconf/db we would issue a warning
once per process -- even if multiple files were missing.

This was enforced using a static variable.

An unfortunate side effect of this global state is that the testcases
couldn't reliably know if to expect the error or not.  This issue was
side-stepped by running any cases that may emit the warning under a
fork, but that made it difficult to debug some of the cases.

Rework the backends not to use global state for the flag and instead
store it per-source.

Remove a use of g_test_trap_fork().