engine: fix a case of variable shadowing
Testing uncovered a case of a shadowed variable in the signal handling
code of the engine: the object path on which a signal arrived was being
shadowed by the 'path' variable used when deserialising the content of
the signal, causing the signal to fail to be delivered.

Rename the D-Bus path variable to 'object_path' to avoid the issue.