engine: fix race in match rule adding
Due to the usual way async in dconf works, we can start getting replies
immediately (without returning to the mainloop).  For this reason, we
must be sure that we do not modify the OutstandingWatch structure at any
point after we've issued even one call because the results may already
be coming in.

For this reason, we must add up the 'pending' count in a separate loop
run before we start making the calls.

https://bugzilla.gnome.org/show_bug.cgi?id=724929