Add support for manually allocating schemas
Add an option "manual-schemas", with type a{ss}, that is inspected
when parsing schemas. For each entry, left-side is a relocatable schema
ID and right-side is a "path specification", i.e. absolute paths that
allow empty segments (//).

The indicated schema is applied to every path in the tree that matches
the path spec in the right-side. Empty path segments (//) function as
wild card in the matching procedure.

The matching procedure matches existing paths (discovered previously
through non-relocatable schemas and DConfClient), but also creates new
paths if the specs are definite enough (i.e. unless it contains some
wild card for which there are none existing paths).

Value example:
{ 'org.gnome.desktop.app-folders.folder':
        '/org/gnome/desktop/app-folders/folders//' }