Make the model dynamic
Every 3 seconds, a new SettingsSchemaSource is created and queried to
detect added or removed schemas. Whenever changes are detected, the
model emits a paths_changed signal indicating the paths that canged.

The direct tree descent (i.e. taking one directory from inside another)
was completely removed from the model together with the
Dictionary.key_model attribute. A new simpler prefix tree now caches
the path information in place of the directories. This structure makes
easier to dynamically handle wildcards in relocatable schemas mappings.
It also does not cache key objects.

Changes in the relocatable schemas mappings mechanisms are also handled
dynamically.

The UI handles the paths_changed signal by checking if the current view
is affected an showing a reload infobar with a warning.

This patch is known not to handle all the cases, see bug and comment at
https://bugzilla.gnome.org/show_bug.cgi?id=731750#c35 for informations.