Rework "need reload" detection
Actually compare the content of the current view with fresh content from
the model. Does not rely on comparing coinciding paths, as that does not
give enough information to decide (notably it is not possible to know
from DConf.Client if a path notified as changed has been created or not).

If the view is stale, either reload directly (when it is known that the
changes are from inside Dconf Editor) or show the "need reload" infobar.

Remove a forced reload in RegistryView when a DConfKey is "ghosted".