Improve request_path signal and handler
Additional parameter "notify_missing" allows suppressing notifications
in case the given path is not found. It is to be set when it is
expected that the given path does not exist, e.g. after deleting a key.

Falling back tries parent-by-parent until some ancestor is found. In
the worst case it falls back to "/".

Keys freshly retrieved from the model are never "ghost", so no need to
deal with that case.