Refactor path initialization and non-strict paths
Only change paths after handling a non-strict path. This is done by
integrating non-strict path handling both in the model and request_path
signal handler and using that facility on initialization.

This fixes the pathbar being confused by setting a non-existing key path
when handling it non-strictly as a folder path (now the path is only set
after resolving the object).