Special case root for search, part 1.
The root path does not have a "local search" functionality, as
that would be similar to the "global search" and so, confusing
as hell. Instead, use an "Open path entry" functionality for a
better discoverability of this function (part 1), and make the
search buttons open for there the search entry empty (part 2).